@@ -16,13 +16,13 @@ android {
             android.signingConfigs.release.keyPassword props.keyAliasPassword
         }
     }*/
-    compileSdkVersion 28
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 29
+    buildToolsVersion '29.0.3'
     compileOptions.encoding = 'ISO-8859-1'
     defaultConfig {
         applicationId "com.wyrnLabs.jotdownthatmovie"
         minSdkVersion 21
-        targetSdkVersion 28
+        targetSdkVersion 29
     }
     /*buildTypes {
         release {
@@ -34,17 +34,17 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:gridlayout-v7:28.0.0'
-    implementation 'com.android.support:support-v4:28.0.0'
-    implementation 'com.android.support:appcompat-v7:28.0.0'
+    implementation 'androidx.gridlayout:gridlayout:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.appcompat:appcompat:1.2.0'
     implementation files('libs/YouTubeAndroidPlayerApi.jar')
-    implementation 'com.android.support:design:28.0.0'
-    implementation 'com.android.support:recyclerview-v7:28.0.0'
+    implementation 'com.google.android.material:material:1.2.1'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
 
     implementation 'com.github.chrisbanes:PhotoView:2.1.3'
-    implementation 'com.github.bumptech.glide:glide:4.2.0'
-    annotationProcessor 'com.github.bumptech.glide:compiler:4.2.0'
-    implementation 'com.android.support:palette-v7:28.0.0'
+    implementation 'com.github.bumptech.glide:glide:4.11.0'
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.11.0'
+    implementation 'androidx.palette:palette:1.0.0'
 
-    implementation 'com.google.code.gson:gson:2.3.1'
+    implementation 'com.google.code.gson:gson:2.8.6'
 }
