@@ -5,8 +5,8 @@ import android.graphics.Bitmap;
 import android.graphics.Color;
 import android.graphics.drawable.BitmapDrawable;
 import android.os.Build;
-import android.support.annotation.Nullable;
-import android.support.v7.graphics.Palette;
+import androidx.annotation.Nullable;
+import androidx.palette.graphics.Palette;
 import android.view.Gravity;
 import android.view.LayoutInflater;
 import android.view.View;
