@@ -18,9 +18,9 @@ package com.wyrnlab.jotdownthatmovie.ExternalLibraries.materialtaptagetprompt;
 
 import android.app.Dialog;
 import android.content.Context;
-import android.support.annotation.IdRes;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
+import androidx.annotation.IdRes;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import android.view.View;
 import android.view.ViewGroup;
 
