@@ -17,7 +17,7 @@
 package com.wyrnlab.jotdownthatmovie.ExternalLibraries.materialtaptagetprompt.extras;
 
 import android.graphics.Color;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.TextPaint;
 import android.text.style.CharacterStyle;
 
