@@ -17,8 +17,8 @@
 package com.wyrnlab.jotdownthatmovie.ExternalLibraries.materialtaptagetprompt.extras;
 
 import android.graphics.Rect;
-import android.support.annotation.ColorInt;
-import android.support.annotation.NonNull;
+import androidx.annotation.ColorInt;
+import androidx.annotation.NonNull;
 
 /**
  * Used to render the prompt background.
