@@ -6,10 +6,10 @@ import android.app.ProgressDialog;
 import android.content.Intent;
 import android.content.res.Configuration;
 import android.os.Bundle;
-import android.support.v4.view.MenuItemCompat;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.ShareActionProvider;
+import androidx.core.view.MenuItemCompat;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.ShareActionProvider;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -20,7 +20,6 @@ import android.widget.TextView;
 
 import com.wyrnlab.jotdownthatmovie.APIS.TheMovieDB.search.AsyncResponse;
 import com.wyrnlab.jotdownthatmovie.APIS.TheMovieDB.search.Movies.SearchInfoMovie;
-import com.wyrnlab.jotdownthatmovie.APIS.TheMovieDB.search.Movies.SearchMovieURLTrailer;
 import com.wyrnlab.jotdownthatmovie.ExternalLibraries.FullImages.PhotoFullPopupWindow;
 import com.wyrnlab.jotdownthatmovie.ExternalLibraries.lazylist.ImageLoader;
 import com.wyrnlab.jotdownthatmovie.Model.AudiovisualInterface;
@@ -32,7 +31,6 @@ import com.wyrnlab.jotdownthatmovie.Utils.ImageHandler;
 import com.wyrnlab.jotdownthatmovie.Utils.MyUtils;
 import com.wyrnlab.jotdownthatmovie.Utils.SetTheLanguages;
 import com.wyrnlab.jotdownthatmovie.View.Activities.SimilarMoviesModal;
-import com.wyrnlab.jotdownthatmovie.View.Activities.YoutubeActivityView;
 import com.wyrnlab.jotdownthatmovie.View.Recyclerviews.RecyclerViewAdapter;
 import com.wyrnlab.jotdownthatmovie.View.TrailerDialog;
 
@@ -52,6 +50,8 @@ public class InfoMovieSearch extends AppCompatActivity implements AsyncResponse 
 	TextView seasons;
 	TextView generoLab;
 	TextView directorLab;
+	TextView originalTitle;
+	TextView originalLanguage;
 	Button botonAnadir;
 	Button botonVolver;
 	Button botonTrailer;
@@ -98,6 +98,8 @@ public class InfoMovieSearch extends AppCompatActivity implements AsyncResponse 
 		botonVolver = (Button)findViewById(R.id.BtnAtras);
 		botonTrailer = (Button)findViewById(R.id.BtnTrailer);
 		botonSimilars = (Button)findViewById(R.id.BtnSimilars);
+		originalTitle = (TextView)findViewById(R.id.OriginalTitleText);
+		originalLanguage = (TextView)findViewById(R.id.OriginalLangugeText);
 
       //Recuperamos la información pasada en el intent
         Bundle bundle = this.getIntent().getExtras();
@@ -205,7 +207,8 @@ public class InfoMovieSearch extends AppCompatActivity implements AsyncResponse 
 
 			}
 		});
-
+		originalLanguage.setText(General.getLanguageTranslations(pelicula.getOriginalLanguage()));
+		originalTitle.setText(pelicula.getTituloOriginal());
 
 		similarMoviesModal.pelicula = this.pelicula;
 	}
