@@ -7,14 +7,14 @@ import android.graphics.Color;
 import android.graphics.PorterDuff;
 import android.graphics.drawable.ColorDrawable;
 import android.graphics.drawable.Drawable;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.widget.RecyclerView;
+import androidx.core.content.ContextCompat;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 import com.wyrnlab.jotdownthatmovie.R;
 import com.wyrnlab.jotdownthatmovie.View.Activities.MainActivity;
 
-public class ItemTouchRemoveHelper extends android.support.v7.widget.helper.ItemTouchHelper.SimpleCallback {
+public class ItemTouchRemoveHelper extends androidx.recyclerview.widget.ItemTouchHelper.SimpleCallback {
     Drawable background;
     Drawable xMark;
     int xMarkMargin;
