@@ -1,7 +1,7 @@
 package com.wyrnlab.jotdownthatmovie.View.Recyclerviews;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.ContextMenu;
 import android.view.View;
 import android.widget.ImageView;
