@@ -1,10 +1,9 @@
 package com.wyrnlab.jotdownthatmovie.View.Recyclerviews;
 
 import android.content.Context;
-import android.inputmethodservice.Keyboard;
 import android.os.Handler;
-import android.support.design.widget.Snackbar;
-import android.support.v7.widget.RecyclerView;
+import com.google.android.material.snackbar.Snackbar;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
